backend/bot/requirements.txt
# Зависимости для Telegram бота и API сервера LVR

# Основные библиотеки
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.7

# Дополнительные зависимости
requests==2.31.0
python-dotenv==1.0.0

# Для работы с асинхронным кодом
asyncio-mqtt==0.16.1

# Логирование и отладка
colorlog==6.8.0
